% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_workspace.r
\name{active_data_info}
\alias{active_data_info}
\title{Helper: get the active microtable info string for display}
\usage{
active_data_info(rv)
}
\arguments{
\item{rv}{Global reactiveValues}
}
\value{
HTML tags with current data info
}
\description{
Helper: get the active microtable info string for display
}
