% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{check_microtable}
\alias{check_microtable}
\title{Check if microtable object exists and is valid}
\usage{
check_microtable(rv)
}
\arguments{
\item{rv}{reactiveValues containing microtable}
}
\value{
TRUE if microtable exists, FALSE otherwise
}
\description{
Check if microtable object exists and is valid
}
