% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{clear_codes}
\alias{clear_codes}
\title{Clear all generated codes}
\usage{
clear_codes(rv)
}
\arguments{
\item{rv}{A reactiveValues object}
}
\value{
Invisible TRUE
}
\description{
Resets the code counter and clears all accumulated codes.
}
