% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{fmt_num}
\alias{fmt_num}
\title{Format a numeric value for display}
\usage{
fmt_num(x, digits = 4)
}
\arguments{
\item{x}{Numeric value}

\item{digits}{Number of decimal places}
}
\value{
Formatted character string
}
\description{
Format a numeric value for display
}
