% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{format_code_download}
\alias{format_code_download}
\title{Download handler for R scripts}
\usage{
format_code_download(code)
}
\arguments{
\item{code}{Character vector of code lines}
}
\value{
Content string for download
}
\description{
Download handler for R scripts
}
