% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{generate_import_code}
\alias{generate_import_code}
\title{Generate R code string for file import}
\usage{
generate_import_code(file_type, paths, object_name = "tmp_microtable")
}
\arguments{
\item{file_type}{Type of import ("qiime2_qza", "qiime2_tsv", "rdata", "manual")}

\item{paths}{Named list of file paths}

\item{object_name}{Name of the resulting microtable object}
}
\value{
Character string of R code
}
\description{
Creates R code for importing various file formats into microtable objects.
}
