% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{get_sample_cols}
\alias{get_sample_cols}
\title{Get available sample metadata columns}
\usage{
get_sample_cols(rv)
}
\arguments{
\item{rv}{reactiveValues containing microtable}
}
\value{
Character vector of column names from sample_table (excluding first if it's IDs)
}
\description{
Get available sample metadata columns
}
