% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{get_tax_ranks}
\alias{get_tax_ranks}
\title{Get available taxonomic ranks}
\usage{
get_tax_ranks(rv)
}
\arguments{
\item{rv}{reactiveValues containing microtable}
}
\value{
Character vector of taxonomic rank column names
}
\description{
Get available taxonomic ranks
}
