% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{info_box}
\alias{info_box}
\title{Create a styled info box}
\usage{
info_box(title, value, icon = "info-circle", color = "primary")
}
\arguments{
\item{title}{Box title}

\item{value}{Box value}

\item{icon}{FontAwesome icon name}

\item{color}{Box color (primary, success, info, warning, danger)}
}
\value{
HTML tags for info box
}
\description{
Create a styled info box
}
