% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abund.r
\name{mod_abund_server}
\alias{mod_abund_server}
\title{Taxonomic Abundance Module Server}
\usage{
mod_abund_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues}
}
\description{
Taxonomic Abundance Module Server
}
