% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_abund.r
\name{mod_abund_ui}
\alias{mod_abund_ui}
\title{Taxonomic Abundance Calculation Module UI}
\usage{
mod_abund_ui(id, lang = "zh")
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Provides interface for calculating and saving taxonomic abundance tables
at different taxonomic ranks using microeco functions.
}
\seealso{
Other community-analysis: 
\code{\link{mod_alpha_ui}()},
\code{\link{mod_beta_ui}()},
\code{\link{mod_composition_ui}()},
\code{\link{mod_core_ui}()},
\code{\link{mod_func_ui}()}
}
\concept{community-analysis}
\keyword{abundance}
\keyword{microbiome}
\keyword{taxonomy}
