% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_core.r
\name{mod_core_ui}
\alias{mod_core_ui}
\title{Core Microbiome Module UI}
\usage{
mod_core_ui(id, lang = "zh")
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Provides interface for core microbiome analysis including identification of
core taxa present in majority of samples and Venn diagram visualization.
}
\seealso{
Other community-analysis: 
\code{\link{mod_abund_ui}()},
\code{\link{mod_alpha_ui}()},
\code{\link{mod_beta_ui}()},
\code{\link{mod_composition_ui}()}
}
\concept{community-analysis}
\keyword{core}
\keyword{microbiome}
\keyword{venn}
