% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_func.r
\name{mod_func_ui}
\alias{mod_func_ui}
\title{Functional Prediction Module UI}
\usage{
mod_func_ui(id, lang = "zh")
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Provides interface for functional prediction analysis including PICRUSt, FAPROTAX,
BugBase, and redundancy analysis.
}
\seealso{
Other community-analysis: 
\code{\link{mod_abund_ui}()},
\code{\link{mod_alpha_ui}()},
\code{\link{mod_beta_ui}()},
\code{\link{mod_composition_ui}()},
\code{\link{mod_core_ui}()}
}
\concept{community-analysis}
\keyword{FAPROTAX}
\keyword{PICRUSt}
\keyword{functional}
\keyword{microbiome}
\keyword{prediction}
