% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_import.r
\name{mod_import_server}
\alias{mod_import_server}
\title{Data Import Module Server}
\usage{
mod_import_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues from app_server}
}
\description{
Data Import Module Server
}
