% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_norm.r
\name{mod_norm_server}
\alias{mod_norm_server}
\title{Normalization Module Server}
\usage{
mod_norm_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues}
}
\description{
Normalization Module Server
}
