% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_norm.r
\name{mod_norm_ui}
\alias{mod_norm_ui}
\title{Normalization Module UI}
\usage{
mod_norm_ui(id, lang = "zh")
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Provides interface for microbiome data normalization/standardization.
Supports multiple methods including rarefy, CLR, GMPR, CSS, TMM, DESeq2, etc.
}
\seealso{
Other data-processing: 
\code{\link{mod_preprocess_ui}()}
}
\concept{data-processing}
\keyword{microbiome}
\keyword{normalization}
