% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_obj_browser.r
\name{mod_obj_browser_server}
\alias{mod_obj_browser_server}
\title{Object Browser Module Server}
\usage{
mod_obj_browser_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues}
}
\description{
Object Browser Module Server
}
