% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_obj_browser.r
\name{mod_obj_browser_ui}
\alias{mod_obj_browser_ui}
\title{Object Browser Module UI}
\usage{
mod_obj_browser_ui(id, lang = "zh")
}
\arguments{
\item{id}{Module ID}
}
\value{
Shiny UI tagList
}
\description{
Provides interface for browsing microtable object properties including
sample information, OTU table, taxonomy table, and basic statistics.
}
\seealso{
Other utility: 
\code{\link{mod_codes_ui}()},
\code{\link{mod_export_ui}()},
\code{\link{mod_workspace_ui}()}
}
\concept{utility}
\keyword{browser}
\keyword{microtable}
\keyword{object}
