% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_preprocess.r
\name{mod_preprocess_server}
\alias{mod_preprocess_server}
\title{Data Preprocessing Module Server}
\usage{
mod_preprocess_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues}
}
\description{
Data Preprocessing Module Server
}
