% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_workspace.r
\name{mod_workspace_server}
\alias{mod_workspace_server}
\title{Workspace Manager Module Server}
\usage{
mod_workspace_server(id, rv)
}
\arguments{
\item{id}{Module ID}

\item{rv}{Global reactiveValues}
}
\description{
Workspace Manager Module Server
}
