% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{read_table_auto}
\alias{read_table_auto}
\title{Read a table file automatically, detecting format}
\usage{
read_table_auto(filepath, force_numeric = FALSE)
}
\arguments{
\item{filepath}{Path to the file}

\item{force_numeric}{Whether to force convert all columns to numeric (default: FALSE)}
}
\value{
A data.frame
}
\description{
Read a table file automatically, detecting format
}
