% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_utils.r
\name{remove_taxa_prefixes}
\alias{remove_taxa_prefixes}
\title{Remove common taxonomy prefixes from tax_table}
\usage{
remove_taxa_prefixes(tax_table)
}
\arguments{
\item{tax_table}{A data.frame containing taxonomic classifications}
}
\value{
A data.frame with prefixes removed
}
\description{
Remove common taxonomy prefixes from tax_table
}
