% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.r
\name{run_app}
\alias{run_app}
\title{Run the microecoshiny application}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to \code{\link[shiny:shinyApp]{shiny::shinyApp()}}}
}
\value{
A shiny application object
}
\description{
Run the microecoshiny application
}
