% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_i18n.r
\docType{data}
\name{translations}
\alias{translations}
\title{Complete translation dictionary}
\format{
An object of class \code{list} of length 2.
}
\usage{
translations
}
\description{
Complete translation dictionary
}
\keyword{datasets}
