% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_workspace.r
\name{workspace_save}
\alias{workspace_save}
\title{Helper: save current microtable to workspace with a new name}
\usage{
workspace_save(rv, name)
}
\arguments{
\item{rv}{Global reactiveValues}

\item{name}{Name for the new workspace entry}
}
\description{
Helper: save current microtable to workspace with a new name
}
